requirements.txt
flask>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
aggdraw>=1.3.0
